shim: Fix race condition about I/O
We need the I/O to be waited for before we try to wait for the
process. If we don't do this, the WaitProcess() call can return
before all the calls to ReadStdout(), ReadStderr and CloseStdin().
This leads to the shim failure (exits) when call into CloseStdin()
for instance. This happens because WaitProcess() delete everything
related to the process from the agent. We cannot expect to do further
calls after WaitProcess() returned.

Fixes #48

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>